# dcag-api-service-java/openapi-run.yaml
# openapi2-run.yaml
swagger: '2.0'
info:
  title: dcag-api-gateway
  description: DCAG API on API Gateway with a Cloud Run backend
  version: 1.0.0
schemes:
  - https
produces:
  - application/json
x-google-endpoints:
  - name: https://dcag-api-service-s4osc3diiq-el.a.run.app
    allowCors: True
x-google-backend:
  address: https://dcag-api-service-s4osc3diiq-el.a.run.app


securityDefinitions:
  firebase:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    # Replace YOUR-PROJECT-ID with your project ID
    x-google-issuer: "https://securetoken.google.com/anz-driver-ops-ritu"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "anz-driver-ops-ritu"


paths:
  /:
    get:
      summary: Hello World
      operationId: root
      responses:
        '200':
          description: A successful response
          schema:
            type: string
    options:
      description: "Cors associated request to root"
      operationId: "root CORS"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /health:
    get:
      summary: Health check
      operationId: health
      responses:
        '200':
          description: A successful response
          schema:
            type: string
    options:
      description: "Cors associated request to health"
      operationId: "health CORS"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /users:
    get:
      security:
        - firebase: [ ]
      summary: "Get all users"
      operationId: "getUsers"
      responses:
        '200':
          description: A successful response
          schema:
            type: array
    post:
      #      security:
      #        - firebase: [ ]
      summary: "Create new user"
      operationId: "createUser"
      responses:
        '200':
          description: A successful response
          schema:
            type: object
    options:
      description: "Cors for createUser"
      operationId: "createUser CORS"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /users/{userId}:
    get:
      security:
        - firebase: [ ]
      summary: "Get user by id"
      operationId: "getUserById"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
        '404':
          description: "User not found"
    put:
      security:
        - firebase: [ ]
      summary: "Update user"
      operationId: "updateUserById"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
      responses:
        '200':
          description: A successful response
          schema:
            type: object
  /users/byPhoneNumber/{phoneNumber}:
    get:
      #      security:
      #        - firebase: [ ]
      summary: "Get user by phone number"
      operationId: "getUserByPhoneNumber"
      parameters:
        - name: phoneNumber
          in: path
          required: true
          type: string
          description: "Phone number of the user"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
        '404':
          description: "User not found"
  /users/{userId}/analytics/events:
    post:
      security:
        - firebase: [ ]
      summary: "Create analytics events of the user"
      operationId: "createAnalyticsEvents"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
        '404':
          description: "User not found"
    options:
      description: "Cors for createAnalyticsEvents"
      operationId: "createAnalyticsEvents CORS"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /users/{userId}/tasks:
    get:
      security:
        - firebase: [ ]
      summary: "Get tasks of the user"
      operationId: "getUserTasks"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: status
          in: path
          required: true
          type: string
          description: "task status such as COMPLETED, IN_PROGRESS"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
        '404':
          description: "User tasks not found"
    options:
      description: "Cors for getUserTasks"
      operationId: "getUserTasks CORS"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: status
          in: path
          required: true
          type: string
          description: "task status such as COMPLETED, IN_PROGRESS"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /users/{userId}/tasks/summary:
    get:
      security:
        - firebase: [ ]
      summary: "Get task summary of the user"
      operationId: "getUserTasksSummary"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
        '404':
          description: "User task summary not found"
  /users/{userId}/tasks/{taskId}:
    get:
      security:
        - firebase: [ ]
      summary: "Get assigned task of the user"
      operationId: "getAssignedUserTask"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
    post:
      security:
        - firebase: [ ]
      summary: "Assign task to the user"
      operationId: "createUserTask"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
    put:
      security:
        - firebase: [ ]
      summary: "Update assigned task of the user"
      operationId: "updateUserTask"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
    delete:
      security:
        - firebase: [ ]
      summary: "Delete assigned task of the user"
      operationId: "deleteUserTask"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
      responses:
        '200':
          description: "A successful response"
    options:
      description: "Cors for createUserTask"
      operationId: "createUserTask CORS"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /users/{userId}/issues:
    post:
      security:
        - firebase: [ ]
      summary: "Create reported issue of the user"
      operationId: "createReportedIssue"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: type
          in: query
          required: false
          type: string
          description: "taskType such as TEXT_TO_AUDIO"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
    options:
      description: "Cors for createReportedIssue"
      operationId: "createReportedIssue CORS"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskType
          in: query
          required: false
          type: string
          description: "taskType such as TEXT_TO_AUDIO"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /tasks:
    get:
      security:
        - firebase: [ ]
      summary: "Get all tasks"
      operationId: "getAllTasks"
      parameters:
        - name: available
          in: query
          required: false
          type: string
          description: "set available flag to true to retrieve only available tasks"
        - name: userId
          in: query
          required: false
          type: string
          description: "logged in userId(TODO: Remove this param once auth is implemented)"
        - name: taskType
          in: query
          required: false
          type: string
          description: "taskType such as TEXT_TO_AUDIO"
      responses:
        '200':
          description: A successful response
          schema:
            type: array
    post:
      security:
        - firebase: [ ]
      summary: "Create new task"
      operationId: "createTask"
      responses:
        '200':
          description: A successful response
          schema:
            type: object
    options:
      description: "Cors associated request to tasks"
      operationId: "createTask CORS"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /tasks/{taskId}:
    get:
      security:
        - firebase: [ ]
      summary: "Get task by id"
      operationId: "getTaskById"
      parameters:
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
        '404':
          description: "Task not found"
    put:
      security:
        - firebase: [ ]
      summary: "Update task"
      operationId: "updateTaskById"
      parameters:
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
      responses:
        '200':
          description: A successful response
          schema:
            type: object
  /users/{userId}/tasks/{taskId}/uploadUrl:
    get:
      security:
        - firebase: [ ]
      summary: "Get upload url for upload image task type"
      operationId: "getUploadURL"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
        - name: fileName
          in: query
          required: true
          type: string
          description: "Image File Name"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
    options:
      description: "Cors for Get upload url for upload image task type"
      operationId: "getUploadURL CORS"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskId
          in: path
          required: true
          type: string
          description: "ID of the task"
        - name: fileName
          in: query
          required: true
          type: string
          description: "Image File Name"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /docs:
    get:
      security:
        - firebase: [ ]
      summary: "Get docs"
      operationId: "getDocs"
      parameters:
        - name: type
          in: query
          required: false
          type: string
          description: "Type of doc"
        - name: language
          in: query
          required: false
          type: string
          description: "Language of Doc"
        - name: name
          in: query
          required: false
          type: string
          description: "Name of Doc"
      responses:
        '200':
          description: A successful response
          schema:
            type: object
    options:
      description: "Cors for createUser"
      operationId: "getDocs CORS"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /ingest-tasks:
    post:
      security:
        - firebase: [ ]
      summary: "Injest tasks"
      operationId: "injestTasks"
      responses:
        '200':
          description: A successful response
          schema:
            type: object
    options:
      description: "Cors for injestTasks"
      operationId: "injestTasks CORS"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /users/{userId}/survey:
    post:
      security:
        - firebase: [ ]
      summary: "Create User survey"
      operationId: "createUserSurvey"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
    options:
      description: "Cors for createUserSurvey"
      operationId: "createUserSurvey CORS"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /users/{userId}/earnings:
    get:
      security:
        - firebase: [ ]
      summary: "Fetch UserEarnings"
      operationId: "getUserEarnings"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: startDate
          in: query
          required: true
          type: string
          description: "Start Date of the completed tasks"
        - name: completionDate
          in: query
          required: true
          type: string
          description: "Completion Date of the completed tasks"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
    options:
      description: "Cors for getUserEarnings"
      operationId: "getUserEarnings CORS"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: startDate
          in: query
          required: true
          type: string
          description: "Start Date of the completed tasks"
        - name: completionDate
          in: query
          required: true
          type: string
          description: "Completion Date of the completed tasks"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
  /users/{userId}/earnings/{taskType}:
    get:
      security:
        - firebase: [ ]
      summary: "Fetch UserEarnings Details by TaskType"
      operationId: "getUserEarningDetails"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskType
          in: path
          required: true
          type: string
          description: "Task type of the user"
        - name: startDate
          in: query
          required: true
          type: string
          description: "Start Date of the completed tasks"
        - name: completionDate
          in: query
          required: true
          type: string
          description: "Completion Date of the completed tasks"
      responses:
        '200':
          description: "A successful response"
          schema:
            type: object
    options:
      description: "Cors for getUserEarningDetails"
      operationId: "getUserEarningDetails CORS"
      parameters:
        - name: userId
          in: path
          required: true
          type: string
          description: "ID of the user"
        - name: taskType
          in: path
          required: true
          type: string
          description: "Task type of the user"
        - name: startDate
          in: query
          required: true
          type: string
          description: "Start Date of the completed tasks"
        - name: completionDate
          in: query
          required: true
          type: string
          description: "Completion Date of the completed tasks"
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"